requirements.txt
pillow
torch
torchvision
numpy
kornia
wandb
tqdm
attrs